added a new feature to delete jobs by passing multi ids, made some tests refactoring

+++ src/main/java/com/epam/grid/engine/provider/log/JobLogProvider.java
@@ -47,6 +47,6 @@ public interface JobLogProvider extends CommandTypeAware {
      * @param logType The type of required log file.
      * @return The job log file like a stream.
      */
-    InputStream getJobLogFile(final int jobId, final JobLogInfo.Type logType);
+    InputStream getJobLogFile(final long jobId, final JobLogInfo.Type logType);
 
 }
